+++ Source/Rows/SliderRow.swift
@@ -64,17 +64,17 @@ open class SliderCell: Cell<Float>, CellType {
             // title
             let title = textLabel
             textLabel?.translatesAutoresizingMaskIntoConstraints = false
-            textLabel?.setContentHuggingPriority(500, for: .horizontal)
+            textLabel?.setContentHuggingPriority(UILayoutPriority(rawValue: 500), for: .horizontal)
             self.titleLabel = title
 
             let value = detailTextLabel
             value?.translatesAutoresizingMaskIntoConstraints = false
-            value?.setContentHuggingPriority(500, for: .horizontal)
+            value?.setContentHuggingPriority(UILayoutPriority(rawValue: 500), for: .horizontal)
             self.valueLabel = value
 
             let slider = UISlider()
             slider.translatesAutoresizingMaskIntoConstraints = false
-            slider.setContentHuggingPriority(500, for: .horizontal)
+            slider.setContentHuggingPriority(UILayoutPriority(rawValue: 500), for: .horizontal)
             self.slider = slider
 
             if shouldShowTitle {
@@ -115,7 +115,7 @@ open class SliderCell: Cell<Float>, CellType {
         contentView.addConstraints(NSLayoutConstraint.constraints(withVisualFormat: "H:|-[slider]-|", options: NSLayoutFormatOptions.alignAllLastBaseline, metrics: metrics, views: views))
     }
 
-    func valueChanged() {
+    @objc func valueChanged() {
         let roundedValue: Float
         let steps = Float(sliderRow.steps)
         if steps > 0 {